feat: remove templates notebook functionality
